## Related Pull Requests & Issues

- Resolve #1271 

Reviewable, only tasks left with @AdelineCelier:

- [x] Ask for new broadcast/subscription Icon and add it to monitor-ui.
- [x] Check if saving should be instant when subscribing/unsubscribing
or once submit button is clicked (current state)
=> UX a bit weird with warning message + banner before saving the form?
- [x] Double-check buttons title

----

- [x] Tests E2E (Cypress)

+++ frontend/src/features/ControlUnit/components/ControlUnitDialog/ControlUnitContactList/index.tsx
@@ -1,33 +1,31 @@
-import { Accent, Button, Icon } from '@mtes-mct/monitor-ui'
-import { useCallback, useMemo, useState } from 'react'
+import { mainWindowActions } from '@features/MainWindow/slice'
+import { addMainWindowBanner } from '@features/MainWindow/useCases/addMainWindowBanner'
+import { Accent, Button, type ControlUnit, Icon, Level } from '@mtes-mct/monitor-ui'
+import { useCallback, useMemo, useRef, useState } from 'react'
 import styled from 'styled-components'
 
 import { INITIAL_CONTROL_UNIT_CONTACT_FORM_VALUES } from './constants'
 import { Form } from './Form'
 import { Item } from './Item'
 import { sortControlUnitContactsByQualifiedName } from './utils'
-import {
-  controlUnitContactsAPI,
-  useCreateControlUnitContactMutation,
-  useUpdateControlUnitContactMutation
-} from '../../../../../api/controlUnitContactsAPI'
+import { controlUnitContactsAPI } from '../../../../../api/controlUnitContactsAPI'
 import { ConfirmationModal } from '../../../../../components/ConfirmationModal'
 import { useAppDispatch } from '../../../../../hooks/useAppDispatch'
 import { FrontendError } from '../../../../../libs/FrontendError'
+import { createOrUpdateControlUnitContact } from '../../../useCases/createOrUpdateControlUnitContact'
 import { Section } from '../shared/Section'
 import { TextareaForm } from '../shared/TextareaForm'
 
 import type { ControlUnitContactFormValues } from './types'
-import type { ControlUnit } from '../../../../../domain/entities/controlUnit'
 
 type ControlUnitContactListProps = {
   controlUnit: ControlUnit.ControlUnit
   onSubmit: (nextControlUnit: ControlUnit.ControlUnit) => any
 }
 export function ControlUnitContactList({ controlUnit, onSubmit }: ControlUnitContactListProps) {
+  const noEmailSubscriptionContactWarningBannerIdRef = useRef<number | undefined>(undefined)
+
   const dispatch = useAppDispatch()
-  const [createControlUnitContact] = useCreateControlUnitContactMutation()
-  const [updateControlUnitContact] = useUpdateControlUnitContactMutation()
 
   const [editedControlUnitContactId, setEditedControlUnitContactId] = useState<number | undefined>(undefined)
   const [isDeletionConfirmationModalOpen, setIsDeletionConfirmationModalOpen] = useState(false)
@@ -60,29 +58,54 @@ export function ControlUnitContactList({ controlUnit, onSubmit }: ControlUnitCon
     setIsNewControlUnitContactFormOpen(false)
   }, [])
 
-  const confirmDeletion = useCallback(async () => {
+  const confirmDeletion = async () => {
     if (!editedControlUnitContactId) {
       throw new FrontendError('`editedControlUnitContactId` is undefined.')
+    }
+
+    // There can only be one email subscription contact per control unit,
+    // meaning that if the user is trying to delete the only email subscription contact,
+    // we need to warn them that the control unit will no longer receive any email
+    const controlUnitContactToDelete = controlUnit.controlUnitContacts.find(
+      controlUnitContact => controlUnitContact.id === editedControlUnitContactId
+    )
+    if (!controlUnitContactToDelete) {
+      throw new FrontendError('`deletedControlUnitContact` is undefined.')
+    }
+    if (controlUnitContactToDelete.isEmailSubscriptionContact) {
+      noEmailSubscriptionContactWarningBannerIdRef.current = showNoEmailSubscriptionContactWarningBanner()
     }
 
     await dispatch(controlUnitContactsAPI.endpoints.deleteControlUnitContact.initiate(editedControlUnitContactId))
 
     closeDialogsAndModals()
     closeForm()
-  }, [closeDialogsAndModals, closeForm, dispatch, editedControlUnitContactId])
+  }
 
-  const createOrUpdateControlUnitContact = useCallback(
-    async (controlUnitContactFormValues: ControlUnitContactFormValues) => {
-      if (isNewControlUnitContactFormOpen) {
-        await createControlUnitContact(controlUnitContactFormValues as ControlUnit.NewControlUnitContactData)
-      } else {
-        await updateControlUnitContact(controlUnitContactFormValues as ControlUnit.ControlUnitContactData)
-      }
+  const submit = async (controlUnitContactFormValues: ControlUnitContactFormValues) => {
+    const hadAnEmailSubscriptionContact = controlUnit.controlUnitContacts.some(
+      controlUnitContact => controlUnitContact.isEmailSubscriptionContact
+    )
+    const willHaveAnEmailSubscriptionContact =
+      controlUnitContactFormValues.isEmailSubscriptionContact ||
+      controlUnit.controlUnitContacts.some(
+        controlUnitContact =>
+          controlUnitContact.id !== controlUnitContactFormValues.id && controlUnitContact.isEmailSubscriptionContact
+      )
 
-      closeForm()
-    },
-    [closeForm, createControlUnitContact, isNewControlUnitContactFormOpen, updateControlUnitContact]
-  )
+    // There can only be one email subscription contact per control unit,
+    // meaning that if the user is trying to unsubscribe the current email subscription contact,
+    // we need to warn them that the control unit will no longer receive any email
+    if (hadAnEmailSubscriptionContact && !willHaveAnEmailSubscriptionContact) {
+      noEmailSubscriptionContactWarningBannerIdRef.current = showNoEmailSubscriptionContactWarningBanner()
+    } else {
+      hideNoEmailSubscriptionContactWarningBannerIfAny()
+    }
+
+    dispatch(createOrUpdateControlUnitContact(controlUnitContactFormValues))
+
+    closeForm()
+  }
 
   const openCreationForm = useCallback(() => {
     setEditedControlUnitContactId(undefined)
@@ -93,6 +116,32 @@ export function ControlUnitContactList({ controlUnit, onSubmit }: ControlUnitCon
     setEditedControlUnitContactId(nextEditedControlUnitResourceId)
     setIsNewControlUnitContactFormOpen(false)
   }, [])
+
+  const hideNoEmailSubscriptionContactWarningBannerIfAny = () => {
+    if (noEmailSubscriptionContactWarningBannerIdRef.current === undefined) {
+      return
+    }
+
+    dispatch(mainWindowActions.removeBanner(noEmailSubscriptionContactWarningBannerIdRef.current))
+
+    noEmailSubscriptionContactWarningBannerIdRef.current = undefined
+  }
+
+  /**
+   * @returns ID of the banner that was added to the main window.
+   */
+  const showNoEmailSubscriptionContactWarningBanner = () =>
+    dispatch(
+      addMainWindowBanner({
+        children:
+          'Cette unité n’a actuellement plus d’adresse de diffusion. Elle ne recevra plus de préavis ni de bilan de ses activités de contrôle.',
+        closingDelay: 115000,
+        isClosable: true,
+        isFixed: true,
+        level: Level.WARNING,
+        withAutomaticClosing: true
+      })
+    )
 
   return (
     <Section>
@@ -108,10 +157,12 @@ export function ControlUnitContactList({ controlUnit, onSubmit }: ControlUnitCon
         {sortedControlUnitContacts.map(controlUnitContact =>
           controlUnitContact.id === editedControlUnitContactId ? (
             <StyledEditionForm
+              key={controlUnitContact.id}
+              controlUnit={controlUnit}
               initialValues={editedControlUnitContact}
               onCancel={closeForm}
               onDelete={askForDeletionConfirmation}
-              onSubmit={createOrUpdateControlUnitContact}
+              onSubmit={submit}
             />
           ) : (
             <Item key={controlUnitContact.id} controlUnitContact={controlUnitContact} onEdit={openEditionForm} />
@@ -120,9 +171,10 @@ export function ControlUnitContactList({ controlUnit, onSubmit }: ControlUnitCon
 
         {isNewControlUnitContactFormOpen ? (
           <StyledCreationForm
+            controlUnit={controlUnit}
             initialValues={editedControlUnitContact}
             onCancel={closeForm}
-            onSubmit={createOrUpdateControlUnitContact}
+            onSubmit={submit}
           />
         ) : (
           <div>
